FashionMNIST dataset to git ignore
