+++ networks.py
@@ -62,3 +62,30 @@ class Critic(nn.Module):
         if self.mode == 'dcgan':
             output = self.sigmoid(output)
         return torch.reshape(output, [-1]), output_vec
+
+
+class InverseGenerator(nn.Module):
+    def __init__(self, latent_dim, model_dim):
+        super(InverseGenerator, self).__init__()
+
+        self.model_dim = model_dim
+
+        def block(in_feat, out_feat, normalize=True):
+            layers = [nn.Conv2d(in_feat, out_feat, kernel_size=4, stride=2, padding=1)]
+            if normalize:
+                layers.append(nn.BatchNorm2d(out_feat))
+            layers.append(nn.LeakyReLU(0.2, inplace=True))
+            return layers
+
+        self.model_body = nn.Sequential(*block(1, self.model_dim, normalize=False),
+                                        *block(self.model_dim, 2*self.model_dim),
+                                        *block(2*self.model_dim, 4*self.model_dim))
+        self.linear = nn.Linear(4*4*4*self.model_dim, latent_dim)
+        self.model_tail = nn.Sequential(nn.Dropout(0.5), nn.Linear(latent_dim, 1))
+        self.sigmoid = nn.Sigmoid()
+
+    def forward(self, img):
+        output = self.model_body(img)
+        output = torch.reshape(output, [-1, 4*4*4*self.model_dim])
+        output = self.linear(output)
+        return output
